// src/lib/api/gen/types/ListCategories.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type ListCategoriesPathParams = {
    /**
     * @type string
    */
    slug: string;
};

/**
 * @description Default Response
*/
export type ListCategories200 = {
    /**
     * @type array
    */
    items: {
        /**
         * @type string
        */
        id: string;
        /**
         * @type string
        */
        slug: string;
        /**
         * @type string
        */
        name: string;
        /**
         * @minLength -9007199254740991
         * @maxLength 9007199254740991
         * @type integer
        */
        position: number;
        /**
         * @minLength -9007199254740991
         * @maxLength 9007199254740991
         * @type integer
        */
        productCount: number;
    }[];
    /**
     * @minLength -9007199254740991
     * @maxLength 9007199254740991
     * @type integer
    */
    total: number;
};

export type ListCategoriesQueryResponse = ListCategories200;

export type ListCategoriesQuery = {
    Response: ListCategories200;
    PathParams: ListCategoriesPathParams;
    Errors: any;
};